% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-attractsegment.R
\name{attract_by_distance}
\alias{attract_by_distance}
\title{Move points towards each other an absolute distance}
\usage{
attract_by_distance(x, xend, y, yend, length_start, length_end)
}
\arguments{
\item{x, xend, y, yend}{The coordinates of the start and end of the segment}

\item{prop_start, prop_end}{How much to take off the start and end of the line (in the units of the data; beware this can make little sense unless both axes have the same units)}
}
\value{
A named list of new coordinates for the start and end of the line
}
\description{
Move points towards each other an absolute distance
}
\keyword{internal}
