% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-attractsegment.R
\name{attract_proportionally}
\alias{attract_proportionally}
\title{Move points towards each other a proportion of the distance between them}
\usage{
attract_proportionally(x, xend, y, yend, prop_start, prop_end)
}
\arguments{
\item{x, xend, y, yend}{The coordinates of the start and end of the segment}

\item{prop_start, prop_end}{How much (proportionally) to take off the start and end of the line}
}
\value{
A named list of new coordinates for the start and end of the line
}
\description{
Move points towards each other a proportion of the distance between them
}
\keyword{internal}
