% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arrowsegment.R
\name{geom_arrowsegment}
\alias{geom_arrowsegment}
\title{Line segments with flexible arrows}
\usage{
geom_arrowsegment(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrows = list(arrow()),
  arrow_fills = NULL,
  arrow_positions = 1,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\description{
`geom_arrowsegment()` is parameterised much as [geom_segment()], and the default
 `geom_arrowsegment()` is equivalent to `geom_segment(arrow = arrow())`. (It is
 assumed that the user wants some kind of arrow.) The extended functionality is to
 allow free placement of the arrowhead anywhere along the segment, and also multiple
 arrowheads.
}
\details{
The function works by dividing the line up into 1 or more segment grobs, each of
 which is generated by [grid::arrow()] except potentially the last (the closest
 to the point (`xend`, `yend'`). The vector `arrow_positions`, whose entries must
 lie between 0 and 1, defines where each arrow segment ends, as a proportional
 position along the line. If the last entry of `arrow_positions` 1, then the last
 grob has an arrow; otherwise it does not.

 The function is designed with the expectation that arrows point from (`x`,`y`) to
 (`xend`,`yend`) but the `arrows` argument will happily accept `arrow(ends = "first")`
 or `arrow(ends = "both")` if you prefer. Just remember that the final segment is
 only an arrow at all if the last entry of `arrow_positions` is 1.

 @eval rd_aesthetics("geom", "arrowsegment")
 @inheritParams layer
 @inheritParams geom_segment
 @param arrow_positions A vector of distinct points on the unit interval. 0 is not
 permitted but arbitrarily small values are; 1 is permitted. The default behaviour is that
 arrowheads will be placed proportionally along the line connecting (`x`,`y`)
 to (`xend`,`yend`) at these points.
 In more detail: The first arrow segment begins at (`x`,`y`) and ends a proportional
 distance along the straight line joining (`x`,`y`) and(`xend`,`yend`) equal to the first
 entry of this vector. The second bridges the first two entries, and so on. If the final
 entry is 1 then the last segment is an arrow (and hence usually an arrowhead will be
 placed at the end of the line). If it is not, then  the last segment is simply a line.
 These will be sorted into order from 0 to 1 if they are not already.
 @param arrows A list of arrows, each generated by [grid::arrow()], defining the arrow
 on the segment ended by the respective entry of `arrow_positions`. If this has length 1
 then all arrows will have the same characteristics. The default is [grid::arrow()]
 with default parameters.
 @param arrow_fills A vector of fill colours for the arrowheads, behaves as
 the `arrow_fill` option in [`geom_segment`].
 @export
 @examples

 # Generate some dummy data

 ten.points <- tibble(line.no = rep(1:5, each = 2), x = runif(10), y = runif(10), position = rep(c("start", "end"), 5))
 five.segments <- ten.points %>% pivot_wider(names_from = position, values_from = c(x,y))

 # Default behaviour

 ggplot(five.segments) +
    geom_point(data = ten.points, aes(x = x, y = y)) +
    geom_arrowsegment(aes(x = x_start, xend = x_end, y = y_start, yend = y_end))

 # Midpoint arrowheads

 ggplot(five.segments) +
    geom_point(data = ten.points, aes(x = x, y = y)) +
    geom_arrowsegment(aes(x = x_start, xend = x_end, y = y_start, yend = y_end),
                      arrow_positions = 0.5)

 Double arrows

 ggplot(five.segments) +
    geom_point(data = ten.points, aes(x = x, y = y)) +
    geom_arrowsegment(aes(x = x_start, xend = x_end, y = y_start, yend = y_end),
                      arrow_positions = c(0.25, 0.75))

 Double arrows, last arrowhead at the end point

 ggplot(five.segments) +
    geom_point(data = ten.points, aes(x = x, y = y)) +
    geom_arrowsegment(aes(x = x_start, xend = x_end, y = y_start, yend = y_end),
                      arrow_positions = c(0.25, 1))

 Double arrowheads of varying appearance and position

 ggplot(five.segments) +
     geom_point(data = ten.points, aes(x = x, y = y)) +
     geom_arrowsegment(aes(x = x_start, xend = x_end, y = y_start, yend = y_end),
                       arrow_positions = c(0.25, 0.75),
                       arrows = list(arrow(angle = 45, type = "closed"), arrow(angle = 25, ends = "both")),
                       arrow_fills = "indianred")
}
