% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arrowsegment.R, R/ggproto.R,
%   R/position-attractsegment.R
\docType{data}
\name{GeomArrowSegment}
\alias{GeomArrowSegment}
\alias{ggarchery-ggproto}
\alias{PositionAttractSegment}
\title{Base ggproto classes for ggarchery}
\description{
Base ggproto classes for ggarchery
}
\seealso{
[`ggplot2::ggplot2-ggproto`]
}
\keyword{datasets}
\keyword{internal}
