% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-arrowsegment.R
\name{split_arrows}
\alias{split_arrows}
\title{Find where to place arrows}
\usage{
split_arrows(x, xend, y, yend, splits)
}
\arguments{
\item{x, xend, y, yend}{The coordinates of the start and end of the segment}

\item{splits}{A vector of points between 0 and 1 determining where on the line between start and end to split the
segment into sub-segments, which will each receive an arrowhead}
}
\value{
A data frame describing the start and end coordinates of the sub-segments. The final segment, ending in 1, is not given (and not given an arrowhead) unless the final element of `splits` is 1.
}
\description{
Find where to place arrows
}
\keyword{internal}
